chore: move DTO classes from service files to Models/ (QUAL-006)
