+++ SysManager/SysManager/Services/EventLogService.cs
@@ -184,14 +184,3 @@ public sealed class EventLogService
         _ => 4
     };
 }
-
-/// <summary>Filter parameters for one query against one log.</summary>
-public sealed class EventLogQueryOptions
-{
-    public string LogName { get; set; } = "System";
-    public List<EventSeverity>? Severities { get; set; }
-    public DateTime? Since { get; set; }
-    public string? ProviderName { get; set; }
-    public int? EventId { get; set; }
-    public int MaxResults { get; set; } = 500;
-}
