+++ SysManager/SysManager/Services/MemoryTestService.cs
@@ -4,6 +4,7 @@
 
 using System.Diagnostics;
 using System.Management;
+using SysManager.Models;
 
 namespace SysManager.Services;
 
@@ -123,14 +124,4 @@ public sealed class MemoryTestService
             return list;
         });
     }
-}
-
-public sealed class MemoryModuleHealth
-{
-    public string Slot { get; set; } = "";
-    public string Manufacturer { get; set; } = "";
-    public double CapacityGB { get; set; }
-    public uint SpeedMHz { get; set; }
-    public uint ConfiguredSpeedMHz { get; set; }
-    public string PartNumber { get; set; } = "";
 }
